Bug#37912472 pb2 failure - router.mrs_metadata_logger_logger_log_level
When the Router shuts down there is a possible race condition between
metadata logging thread start and stop. This is more likely with a very
slow runtime (like Valgrind). The stop() is called from the mysql_rest_service
plugin's stop() method, while there is schema_monitor still running that can
call start(). This can lead to the destruction of an unjoined std::thread
that results in crash.

This patch changes the order in the plugin's stop: the logging thread stop
is done after the schema_monitor has been stopped.

Change-Id: Ie6be2b436a247b0d4eb3be6c2e76015d97eae68a