Bug#37784391 - Fix failing mtr tests when running                with --hypergraph option [engines]
nist_all:
A few queries needed sorted_result and few queries were generating
different warnings with hypergraph. Other than these, one of the query
passes the expectation from nist suite with hypergraph where as the
old optimizer does not. Created a new result file because of that.

insert_decimal:
Some of the queries give different result with hypergraph because of
constant folding E.g. where col <> 99999.99999 is rounded of to
col <> 1000000 for a col of type float (10,5) which is later converted to
always true by HG (invokes const folding again). So a value "100000.00000"
is returned for HG and not so in OO. So HG might look correct here.
However if the same condition was col < 99999.99999 and this gets converted
to col < 100000 for OO and always true for HG, then OO would return the
correct result and not HG. This is observed for only deprecated datatypes and
also falls within the scope of the documented problems with comaprisons
in floating point types. So a different result file is created for HG.

type_bit_iuds:
Many queries in the test file give different results with hypergraph. But
the problem is not specific to HG. It just chooses a different plan to
that of the old optimizer. The same problem can be seen in old optimizer
if a query chooses range scan. The main difference in the results come
because of const optimization being used for the old optimizer. Const
optimization ignores the errors while storing the "bit" type data where as
range optimizer handles the error and concludes it as an impossible
range leading to different results. A bug is already present for this
difference in behavior - Bug#36070629.
So we disable the test for hypergraph for now. We can enable it when the
bug gets fixed.

insert_time:
Hypergraph chooses a different plan which results in correct result whereas
old optimizer is choosing index condition pushdown  which is not supported
by hypergraph. Old optimizer is currently returning wrong result. So
we skip the few statements when running with hypergraph.

Change-Id: I07e914e13ec902583c0e6bc5021a57d9f57a6018