Bug#36330455 : long semaphore wait crash after failed MEB incremental                backup couldn't unregister redo consumer
Description:
------------
When there is a redo log consumer lagging behind,
the redo logs that are old can not be remomved
leading to the redo log writer thread waiting for
the consumer to advance. During this there happens
a long semaphore wait crash.

Analysis:
---------
When there are lots of redo logs being generated
and there is a lagging consumer, a situation can
occur where dict_stats_thread holds an x-latch
on dict_operation_lock while waiting for space in
the redo log and srv_master_thread also tries to
acquire x-latch on dict_operation_lock to perform
eviction from the table cache. This leads to long
semaphore wait. This is because log_free_check
method that checks for free space in redo log
relies on the free_check_limit_lsn value that is
calculated based on the checkpoint LSN value and
the capacity. This is not correct in this case,
because the redo logs that are yet to be consumed
by lagging consumer are not considered in this
calculation.

Fix:
----
* Modify the calculation of free_check_limit_lsn
  to consider the oldest consumer's LSN instead of
  the checkpoint LSN.
* Modified the consumption_requested() of the
  clone's archiver consumer to abort if it is
  lagging for 2 seconds.
* Removed log_request_checkpoint_in_next_file and
  log_request_checkpoint_in_next_file_low methods.

Change-Id: I2d4428cf49c89481a9874649f0469c027e57cc6b