Bug#37952274: expr BETWEEN expr and expr fails: Assertion failed
The problem is an assert for unhandled error after resolving
a BETWEEN clause. The BETWEEN clause has two string arguments as
argument 2 and 3, and these are evaluated during resolving to see
if they are equal. One of those evaluations causes an error,
but the result error is not handled properly.

The fix is to handle the error during resolving.

Change-Id: I5fc7f40a60c0edb541f1904d6438e91875d9b12d