WL#16888  Change built-in server default for back_log to 10000
Description:
  This worklog changes the default value of the system variable back_log
to 10000. This change takes effect when this variable is not configured
(either via command line or via configuration file) or explicitly set to
default via SET PERSIST_ONLY

Note:
  The previous default and the minimum value was 0. 0 was used as a
placeholder to udpate the value to `max_connections` during bootstrap.
As current default is updated to 10000, this section is removed. Also,
the minimum value of the `back_log` is set to 1 as 0 is no longer
required.

Change-Id: I3e788c0875121a868d319b2f60da085df2f5e599