Bug#37235041 - Duplicate Detection Flaw for NULL Values in UK During DDL Rollback Operations
Issue:
When a delete operation is performed while the unique key index
is being built, the operation is queued up and applied later
just before the build completion. An insert operation
performed in parallel with the same key values(as the deleted
key), which include a NULL value succeeds. Now if the delete
operation is rolled back, it tries to do a insert and reports
a key violation because of the insert.

Cause:
Unique keys with NULL values are being considered as duplicate
during the rollback back of a delete operation which was
performed in parallel with the index build.

Fix:
If the value of the unique key is NULL skip marking it as a
duplicate.

Change-Id: I46654ff15548d989cbd6b90ae41697bd86536774