Bug# 37451177 - sys.diagnostics() returns "Access denied" error
Problem:

Continuation of previous bug 34189183. The fix for previous bug only
enabled the 'current' config for the diagnostics procedure. However,
it did not fix the usage of 'medium' or 'full' config.

Root Cause:

The diagnostics procedure also uses ps_setup_reload_saved.sql and
ps_setup_save.sql when called with 'full' or 'medium' config.

The SQL over there uses the "SET sql_log_bin" to enable or disable the
binary logging, to avoid the replication of temporary tables. The
privileges required to do that are not available to the root user,
hence getting access denied.

Solution:

Wrap the enable or disable of binlog in a conditional IF statement, so
that the execution of the "SET sql_log_bin" statement is skipped.
Just like the way it was done for bug 34189183.

Change-Id: I69cb9b0337fd9c9102802c93bc4096c960396e0b