Bug#37498680 mysqldump --routines doesn't work with servers < 9.2.0
Skip the libraries dump if the I_S.LIBRARIES view does not exist.
Post-push fix for DD checksum - undo the shuffling the view registration because it modifies the DD checksum.
Use a boolean variable instead. NOTE: This variable will be compiled out in release builds, not impacting the performance a bit.
Remove bug# reference from file names and test output.

Change-Id: I171791d9d5e2ac61e4660efb46c3e2b569cdede9