Bug#37784391: Fix failing mtr tests when running with --hypergraph option [engines, myisam, noclose]
The engines/iuds.delete_year test fails with a result difference when
running with the hypergraph optimizer enabled and default storage
engine set to MyISAM.

The difference is caused by DELETE statements that are
non-deterministic. They use a LIMIT 2 clause to limit the number of
rows that are deleted, but they have no ORDER BY clause, so it depends
on the query plan which two rows are seen first and get deleted.

Fixed by adding ORDER BY clauses to make the DELETE statements
deterministic.

Change-Id: Id384d620c43d13be67eccac1d3e2b61accbf9187