Bug#37989752: AfterSelect hook not called for external tables
notify_after_select is not getting triggered when query contains both
innodb and external tables. This, may in-turn, result into other issues
in secondary_engine management, if defined.

To fix this issue, we ignore the primary engine optimality check for
external tables and ensure that notify_after_select is called if an
external table participates in the query.

Change-Id: I1c8e48b5e019d8e7276ddcf6f67b7a79a610fdb7