WL#16940  Log available Server resources at Startup
Description:
  The MySQL server uses different methods to identify the total number
of logical CPUs and total physical memory it can access. This worklog
allows the MySQL server to log this information at startup. This change
is seen only during normal startup i.e., when neither of `help`,
`validate-config`, `initialize` or `initialize-insecure` is passed. This
information is printed after the PID of the process is logged.

Change-Id: Ifb34a6c867f6a957f8ce2d7c66ce051d2cd35409