WL#16938  Startup option server_memory
Description:
  The MySQL Server auto-tunes the default values of certain
configuration parameters based on the total physical memory available to
the server. This information may come from the cgroup configuration or
from the system hardware configuration.
  This worklog allows the user to configure the amount of memory
considered by the server when determining the default values. It adds a
system variable called server_memory as a read_only non_persistent
variable with a default value of 0 indiciating no limits set.
  In case the server is configured to use more memory than what the
server can access, the server throws an error during bootstrap and
halts.

Note:
  If the value is not provided, server_memory will default to 0. When
the value is 0 (either default or manually set to 0), the server will
fallback to available physical memory using cgroup or hardware as
implemented in WL#16484.

Change-Id: I39310955d952fef2b65a767c74481660daef4f2c