WL#15413 Subscriber row-hash binlog filtering (NdbApi) [2/6]
Problem:

The API node subscriber that wants to only receive a slice of data events from
the data nodes must be able to configure its subscription to SUMA
events.

Analysis:

Subscription options  are generally set through the
`NdbEventOperation`, where it packs information to be sent through the
`SubStartReq`. SUMA expects subscriber slice configuration to be sent
on the `requestInfo` field of `NdbEventOperation`, using the bit
arrangement defined on the `SubStartReq` signal structure.

Solution:

This commit implements the NdbApi functionality to provide
configurable redundancy level for API subscribing nodes. A new
`NdbEventOperation` function sets the necessary bits to enable row
slice configuration; accepting the `count`, that defines into many
slices is the epoch split into, and the `id` of the row slice the
subscriber wants to receive on the event stream.

Change-Id: I93e7b3067f56ec4f3f58832b2275ff61a19e7c90