WL#15413 Subscriber row-hash binlog filtering (NDB plugin) [3/6]
Problem:

For the MySQL NDB Cluster binlog server to only subscribe to a slice
of the row data events, it needs the NDB binlog thread to be able to
configure it and user facing options to specify the configuration.

Analysis:

The MySQL NDB binlog server uses the NdbApi to configure subscription
to SUMA events, by parameterizing the `NdbEventOperation`. The
`NdbEventOperation` function, to configure the row slice of the
subscription, is available to be used.

But, since it will be the end user of the data before being written to
replication file (binlog file), it needs to deal with what is
supported. What is *not* supported is what cannot be split, due to the
nature of splitting the epoch data across different streams:

- System tables (ndb_apply_status, ndb_schema, ndb_schema_result, etc)
- Tables with BLOB columns (1)

(1): NDB Tables with BLOBs cannot have row slices, because they are
implemented as a split of data between the main table and a parts
table (NDB$BLOB_x_y). The parts table individual rows with would, most
likely, be distributed to a different subscriber set and thus result in
fragmented data.

Solution:

This commit adds two new MySQL server parameters to configure row
slices for the MySQL NDB API node:

* `ndb_log_row_slice_count`: the number of row slices that exist in
this server's set of Binlog servers
* `ndb_log_row_slice_id`: the slice id this server binlogs

The options are readonly, so a restart of the server is needed to
change the configuration.

For system tables, the Binlog thread will ignore any row slice
configuration.

For unsupported tables, the Binlog thread will reject the
configuration and fail the setup, to prevent further problems
downstream when event data is incorrectly binlogged.

Change-Id: I548f8e616b6f261105150c76532fdec50d8c417f