Bug#37993516  Handle cgroup v1 default memory limit
Description:
  The MySQL Server discovers the amount of physical memory and total
number of logical CPUs it can access. This is used to calculate the
default values of its configuration parameters enabling it to auto-tune
the defaults based on the system.
  This information may come from the cgroup limits or the system
hardware. The server discovers the cgroup limits by reading the
respective config files.
  In cgroup v1, the memory limit of `LONG_MAX - (LONG_MAX % PAGE_SIZE)`
is written to the config file by default. In cgroup v2, the string "max"
is written to the config file by default.

Issue:
  The APIs used to fetch this information did not account for the
default memory limits configured by cgroup v1.

Fix:
  The APIs are modified to report 0 as the memory limit if the default
memory limit is read from the cgroup config file. It returns 0 to be
consistent with the design choice of using the value of 0 to indicate no
limits being set.

Note:
  The cgroup v2 is not affected as the APIs parse the contents of the
text into an unsigned integer field. This parse fails when reading the
string "max", thus reporting no limits being set.

Change-Id: I67fa1a7d06c5a2e9ac5e53d9516f6ac3b037ea66