Bug#38004868: Reading data from temporary tables in 9.3.1 with               big_tables/hypergraph_optimizer crashes MySQL
A query which sorts on row-ids on top of a hash outer join whose probe
table is empty results in server exit when temp tables needed by the
query are created on disk. Problem arises only with hypergraph
optimizer as the old optimizer has a materialization step below the
row-id sort. When a probe table in a left hash outer join is empty,
build tables are never accessed (which is an optimization introduced
as part of the bugfix for Bug#34940000). For this case, it makes the
executor not open the innodb temp table which is on the build side.
However, the sorting iterator which in this case is sorting on
row-ids, tries to initialize the temp table which is not opened yet
leading to problems.

Solution is to create a Zero Rows iterator when filesort does not
find any rows ro sort.

Change-Id: I58254b5e687ca116106837be1d9bb264b9cc6af3