Bug#37621360: Importing Table with Instant Drop Column fails
Problem:
--------
When importing a table with a secondary index after an instant
column drop there was a failure due to using incorrect fields
mapping both in index rec_cache and in array of fields.

Fix:
----
1. Array of fields used by secondary index is rebuilt during table
import with column drop to make sure it reflects actual list of
fields after an instant drop.
2. rec_cache is set to empty after instant column drop because
it is used only for intrinsic tables.

Change-Id: I66d2c02f90f95205cc7f28b5f036656a256f15ad