Bug#38020526  Set back_log to max_connections if 0
Description:
  WL#16888 had set the default value of back_log to 10000. It also set
the minimum value of back_log to 1 and removed the section which updates
the value to max_connections if the value was 0 (the previous default).

Issue:
  This introduces an unexpected change in behavior where back_log would
take the value of 1 if the user had explicitly configured it to 0. This
causes performance regressions as the value 1 is too small and leads to
frequent reconnects.

Fix:
  Allow 0 to be the minimum value and update the value of back_log to
max_connections if back_log takes the value 0. Ensure that by default
back_log still takes the value 10000.

Change-Id: Idc849c1bb73cf824f90fb8a5f2811cc8c54d586f