Bug#36342348 - MVCC is broken for small partial update of blob
Problem:

When there is a mix of partial update of lob (which provides MVCC
in lob) and the small partial update of lob (which provides MVCC
using undo), the current attempt to build an older version of the
LOB doesn't work.

Let the latest version of LOB be Y and the version of the LOB we
need is X.  Currently we follow these steps:

1.  Obtain version X from the lob module. The idea was this will
    obtain the correct version of w.r.t the partial update feature.
2.  Apply the undo logs, and this will get the correct version
    w.r.t the small partial update feature.

But this doesn't work as expected, demonstrated by the test case
scenario, where the changes done by partial update and the small
partial update overlap each other.

Solution:

We currently solve this issue by restricting the small partial
updates to version 1 of the blob.  This will avoid interaction
between small partial updates and the partial updates.  For the
scope of this bug, this would be sufficient.

Thanks to Fungo Wang <bo@fungo.me> of Alibaba for identifying the
issue with a test case.

Change-Id: I30add96af39ba3f3844939c5e35d2066f9921071