BUG#38042469 InvalidFree(void*) () from /lib64/libtcmalloc.so.4
When router was linked with libtcmalloc 2.6.1, it would crash at exit
with: Attempt to free invalid pointer.

Crash was caused by https://bugzilla.redhat.com/show_bug.cgi?id=1494309,
SecureMemoryPool was using std::aligned_alloc() to allocate aligned
memory blocks, which caused a crash when memory was freed.

The implementation was updated to use internal memory functions which
deal with the allocation of aligned memory.

Change-Id: I939b5448b7816bf2618581fc8bc40f02ac5157fd