Bug#35912852 CAN'T TRUNCATE PARTITION WHEN PARTITION_ID EXCEED INT_MAX
Problem
-------
TRUNCATE PARTITION operations could fail with duplicate key error

```
"ERROR 1062 (23000): Duplicate entry '2147483648-0-0' for key
'table_partition_values.PRIMARY'"
```

when the `partition_id` in the `mysql.table_partitions` reaches the
INT_MAX value.

Analysis
--------
In general TRUNCATE PARTITION works like any other DDL where a new
`dd:Table` object will be persisted in the DD once the SE operation is
done.

During persistence it recursively updates all the children tables like
`indexes`,`foreign_keys`,`partitions`,`triggers`and `check_constraints`.

While persisting `partitions` to DD, it performs insert/update/delete
operations on the `mysql.table_partition_values` DD table.

To check whether we to insert or update the
`mysql.table_partition_values` DD table, the server creates a primary
key from the partition_id and performs search on the table. If the entry
is found, it performs an update, else it performs the insert.

However, when the partition_id is >= INT_MAX, the partition_id
momentarily undergoes integer overflow due to the ulonglong->int
conversion by the constructor of `Table_partition_values_pk`.
Because of this, the partition_id of value `2147483648` will be searched
in the table as `-2147483648`, causing the SE to return as
ROW_NOT_FOUND, thereby server performing the insert on the DD Table
resulting in the query failing with ER_DUPLICATE_KEY

Solution
--------
- Use the typedef `Object_id` (unsigned long long) as the data type
  of the object_id while creating primary key using
  `Table_partition_values_pk` to prevent integer overflow.
- Replace the hardcoded ulonglong to `Object_id` for type `Composite_pk`

Change-Id: I80dafaa2b3b51f5555b943ed16f08b136cf06760