Bug#38005202 Server_memory ignored [1/2]
Description:
  The default values of some configurations are auto-tuned based on the
total physical memory available to the server. WL#16938 introduced the
startup option server_memory to allow the user to configure the amount
of memory to be considered when auto-tuning the defaults. These APIs
are exposed to the server code using the library_mysys component.

Issue:
  Due to the design of the library_mysys component, the APIs were
exposed as STATIC library. Thus, components which link library_mysys
bring their own copy of the APIs. These copies are not confiugred based
on the value of server_memory. This leads to server_memory configuration
being ignored and total_physical_memory() being returned instead.

Fix:
  During initialization of these components, read the system variable
server_memory. Then configure the APIs based on the value read. At any
point if there is failure, throw a warning about the failure and fall
back to the total physical memory available.

Change-Id: I5550bc14bcc424a95b10fec58c0968a39e949584