Bug#37602657 virtual index corruption
Problem:
In some scenarios fields in secondary indexes based on virtual columns
become corrupted with random or invalid values.

Cause:
When an inplace update happens to a secondary index keyed on an virtual
column, the format of the update vector is confused - the field_no
member contains the actual field number, but some code wants it
to be the virtual column number.

The alternative meaning of field_no makes sense for clustered index,
which does not have fields for virtual columns. When for various
reasons we need the update vector to contain information about updates
to virtual column values for the row, we create "virtual field updates"
which do not actually update any fields.

Fix:
"Virtual field updates" with alternative semantics for field_no should
only be used for clustered index update vector. For updates to
materialized virtual column values in secondary indexes, we treat them
as regular field updates.

Change-Id: I6574d5df9ad0f6c527a8e020cb1eb44a87f88f2f