Bug#37617712 routing does not restart after DNS failure
When router fails to resolve the destination adresses of the
cluster-nodes due to a temporary DNS outage, it closes the acceptor
sockets but sporadically fails to open the acceptor socket again.

Background
==========

If DNS comes back without 10 seconds,

1. the metadata-cache opens the acceptor socket again
2. the routing-connect waits for the PRIMARY failover for 10 seconds and
   closes the acceptor again (without anyone ever opening them again).

Resolver errors should not lead to a "wait-for-PRIMARY-failover" as the
wait should only be done if the connection to the primary failed.

In case of DNS failures the PRIMARY may be up and running without any
intention to every switch to another node.

Change
======

- handle resolver errors correctly to not lead to a wait-for-primary-failover

Change-Id: I11ff612063fa53333ad29c7ccdbaea11b9123dc1