Bug#37978404 Schema change not propagated in case of active remote client
Problem

Server table_share invalidation vs schema change race condition results
in inconsistent schema views across MySQLDs.

Scenario

A participant MySQLD has a local session holding a Metadata lock (MDL)
which conflicts with an MDL required to participate in schema
distribution.
The Ndb Binlog Injector (BI) invalidates the Server's table_share,
and then waits for the MDL to become available.
If the session then attempts to access the table again, it can create
a new table_share, referring to the old schema.
If the BI later obtains the MDL and updates the DD then the Server
is left with :
  - New Ndb table
  - New DD Content
  - Old table_share state

This can cause further problems if further DDLs are executed based on
this hybrid state.

The ndb_ddl_open_trans and ndb_rpl_ddl_open_trans testcases are
extended to trigger these situations with :
 - Inplace ALTER add column @ Server X, MDL @ Server Y
 - Inplace ALTER add index @ Server X, MDL @ Server Y
 - Inplace ALTER add column @ Server Y, MDL @ Server X
 - Inplace ALTER add index @ Server Y, MDL @ Server X

This results in schema view divergence, and eventually an
assertion failure due to divergence between the Server DD + NdbApi
views.

ha_ndbcluster is improved with extra self-checking of alignment
between DD and table_share (DEBUG only).

Extra table_share invalidation is added :
 - Under protection of MDL Exclusive lock
 - Executed after commit of MySQL DD schema transaction

This should ensure that
 - There are no concurrent sessions accessing the table
   (due to the MDL Exclusive lock)
 - The old table_share is discarded
 - Any concurrent sessions attempting to create a new
   table_share can only do so after the commit of the
   DD changes, so will use the new schema content.

Change-Id: I7af2cff2929e07d07f84c12df4c2596c23dd9823