Bug#34496831 Move update of "the other checkpoint header" to log_start from recv_recovery_from_checkpoint_start()
Currently, recv_recovery_from_checkpoint_start() updates the other
checkpoint header to ensure the only copy of it is not lost. There are
two locations of checkpoint and the location to which the checkpoint_lsn
is written to is alternated every time. The
recv_recovery_from_checkpoint_start() tries to keep both the values same
during recovery, but it results in writing to the REDO log files during
recovery.

This patch addresses this by having log_start() pick which of the two
checkpoint locations need to be written next so that the "other"
location is chosen first. Using this, the earlier guarantee around how
long the checkpoint_lsn value is seen during recovery is still same
without doing REDO log writes during recovery.

Change-Id: Ib211972e0db3c0433347136a7074e0e88169fe84