Bug#37847144: MySQL #8: an assertion failure
This is a regression from the fix made for Bug#35855294.
That patch made changes to handle correlated subqueries
with window functions. As part of the fix, once a
replacement field (in the window output table) is found
which needs to be used in place of the outer reference
in the correlated subquery during evaluation, the
replacement field is marked as an outer reference. While
marking it as outer reference, code always presumed the
underlying expression for the outer reference is a column
from table. However, for the failing query its a view
reference (field from a merged derived table) pointing
to a function.
We fix it by choosing to update the table reference for
the replacement field only when the underlying expression
of the outer reference is a column. If it's not a column,
then a reference to the replacement field is created with
the outer reference information and we use the reference
as the replacement. This is necessary because used_tables()
for an Item_field will first check for a table reference.
If there is none, it treats it as a field from temporary
table. However, in this case we want it to be treated as
a outer reference. Hence we use an Item_ref instead of
an Item_field as replacement.

Change-Id: Ia08895e7dee0cd3e1f86ff7e0b864ecfb1e678cc