Bug#38094428  innodb_dirty_pages_at_shutdown fails intermittently
Description:
  The test simulates a scenario of a crash during the shutdown command
execution. It expects that the shutdown command would crash the server
and then it starts the server after checking for the corresponding error
log message.
  In the test, the script shutdown_mysqld.inc sets a timeout of 60
seconds and issues the shutdown_server MTR command. The test runner
queries the PID of the server. Then it waits until the server process is
no longer active. However, if this wait exceeds the timeout duration,
the command is reported as failed (with error=2).

Issue:
  When the MTR issues the shutdown_server command, the mysqld server
crashes due to the bug scenario. When the process crashes, the windows
machines sometimes triggers the JIT debugger. This debugger creates a
popup requiring manual intervention. This popup prevents the process
from becoming inactive. Thus, the shutdown_server command times out and
the test fails.
  When ASAN is turned ON, it catches memory allocations from the
Perl_safesysmalloc called in libperl.so library. When the server
crashes, these memory allocations are not freed. ASAN reports this
leaks, but they are not caused by the server.

Fix:
  Skip running the test on Windows and when ASAN is turned ON.

Change-Id: If9dce5c60e5775f6fe35c0a4eef303b79c4514d8