Bug#38095630: Assertion `hour != 0 || minute |= 0 || second != 0'
In the new implementation for TIME, we flag that a negative zero is
not allowed, and a generated negative zero is always converted to
positive zero. However, there was som code that did not adjust this
value correctly.

The fix adds this adjustment where needed, as well as extending the
is_valid() check to trigger on the special negative zero value.

Change-Id: I6f50ec1c91162a7e3c8f516a670d7764736b90df