Bug#38098565 routertest_component_gr_notifications fails on pb2
Router log contains this unexpected error:
// ERROR: 1064 executing statement failed: expected metadata_schema_version
to be a array, got undefined
This leads to the metadata-cache reinitializing additional time which
the test does not like.

The problem is with the mysqld.global.metadata_schema_version mock
server global that is set to the valid value in the global scope
as in other trace-files, but the difference in this particular trace
file is, that it calls prepare_statement_responses() not in the same
global scope but per statement. This sometimes leads to the
metadata_schema_version being reset to `undefined` when other variables
are set from the test.

This test moves initializing the global variables from the global scope
to the per-statement function where prepare_statement_responses() is
also called.

Change-Id: I5b231d35b80b7d7a1998b15cc3b5687afbf20912