Bug#38106166 routertest_component_socket_close fails sporadically
Running

  ./runtime_output_directory/routertest_component_socket_close \
  --gtest_repeat=100 \
  --gtest_filter=*FirstAvailableDestMetadataCacheTest/first_available_gr_v2

sporadically fails with:

  mysql/router/tests/component/test_socket_close.cc:1969: Failure
  Value of: wait_for_port_unused(*router_rw_port)
    Actual: false
  Expected: true

The test expects the RO and RW ports to be closed, but they are actually
open.

Background
==========

The failure happens when connection to the PRIMARY (which is dead) fails
and metadata cache connects to the SECONDARYs which report that the
cluster is fine.

That failover isn't supposed to happen at this point of the test.

Change
======

The problem is that simulate_cluster_node_down receives all the cluster
nodes which allows the fail-over to happen. The simplest fix would be:

  -  simulate_cluster_node_down(node_ports, node_http_ports[0]);
  +  simulate_cluster_node_down({node_ports[0]}, node_http_ports[0]);

To prevent this oversight in the future, the fix also renames some
variables and sections.

Change-Id: Ibdb6c3490836042803ad8073678d0e9ae2e7b627