Bug#38106593 routertest_component_read_replica fails on pb2 ubuntu_ASAN
The test if failing because there is a missing log entry when the Router
could not connect to any of the metadata servers.

The condition in the Router for printing the missing log is invalid,
wrong flag is checked: `!terminated` instead of `!md_servers_reachable`

This patch uses the proper flag and removes the `terminated` parameter
from the API as it is never used now.

Change-Id: I8d98e1097421fca7e5b2bf6566650e28b63e3e91