Bug#37826935 Assertion `semijoin_group.size() == CountOrderElements(path->sort().order)' failed.
This commit fixes the above bug by removing the failing assert. The assert
was failing because LogicalOrderings::ReduceOrdering() treats (sort)orderings
and groupings differently. In orderings, elements keep their original order,
but for orderings they may be reshuffled by
LogicalOrderings::RecanonicalizeGroupings(). And ReduceOrdering() removes
an element if it is functionally dependent on an earlier element. Therefore,
an ordering may provide a given grouping even if they do not have the same
number of elements.

Change-Id: Ifa4e58c0ce681c997b56f7323d125ab85f9beead