Bug#38111817 routertest_integration_routing_sharing fails sporadically
The "change_user" sockets tests sporadically fails.

  $ routertest_integration_routing_sharing \
    --gtest_filter=*.change_user*

Background
==========

The test's SetUp() waits 10 seconds for the stashed-connections to become 3,
but in reality it is 4.

After waiting 5 seconds, the connection-pool closes the unused stashed
connections which either changes the stashed-connections from 4-to-3 or
from 4-to-2-or-less.

In the 4-to-3 case, the test continues and passes, but took 5 seconds.

In the 4-to-2-or-less case, the test fails.

The 4th connection comes from connection-pool which depends on the other
tests run before.

Change
======

- ensure that the connection-pool is cleared between each run to get a
  stable count of connections.

Change-Id: I05e29ee9d091fbecd5150c7764df1faba631f270