Bug#38111885 MySQLSession object is leaked when wait_gtid times out on RO node.
This issue can be visible when running the following test with ASAN:
./mtr router.metadata_cluster_basic_single_primary_sync_gtid --sanitize --record
The sanitize_report shows all sorts of memory leaks related to mysql session.

While adding more traces one can see that there is a session object
leaked after the SELECT 0=WAIT_FOR_EXECUTED_GTID_SET(?, ?) query
times out.

The problematic code is around the handling of `wait_gtid_executed` in the
Router mysql_rest_service plugin. The logic there is so that if
WAIT_FOR_EXECUTED_GTID_SET times out on a RO node, the RO session that was
used to run this query is overwritten with a RW session from the cache.
The problem is that the CachedObject::operator=() does not
release currently assigned session when overwriting with a new one.

This patch fixes the CachedObject::operator=() to release the old session
object before assigning a new one.

Change-Id: I9ab75cd7eef1e581d9812fe0cc3836c7ef88b67b