Bug#37453240 Run include-what-you-use and fix warnings [vio,noclose]
CC=clang CXX=clang++ cmake <path to src> -DWITH_IWYU=ON

In addition to fixing warnings for vio, this patch introduces a '.imp'
file for general include-what-you-use rules.

The patch implements a new cmake option WITH_IWYU, which is OFF by
default. If set to ON, it will do:
  SET(CMAKE_CXX_INCLUDE_WHAT_YOU_USE
    include-what-you-use
    -Xiwyu --mapping_file=${CMAKE_SOURCE_DIR}/iwyu_mappings.imp

Note that this assumes that include-what-you-use is in your PATH, and
that it works only for clang, not gcc.

Change-Id: I0d7da222924eeae6534fd28b0785de3253599637