Bug#38122221 routertest_component_gr_notifications/AddNode fails randomly on pb2
The test sometimes fails on the following check:

Expected equality of these values:
  1
  count_str_occurences(log_content, needle)
Which is: 2

The problem is triggered by the issue seen in the log:
...
  While updating metadata, could not establish a connection to cluster
    'test' through undefined:11080
..

The test sets the mock globals not setting one of them "gr_node_host", relying
on it being set to default value in the js trace file if it is undefined.
The problem is that it can be reset back to undefined between it was
set to default and the Router does a query that uses it.

This patch makes the test to always set "gr_node_host" when instrumenting
metadata in the mock server.

Change-Id: Ie8090d679ad1542d70da95bf0a5e168236707483