Bug#38136428 Set cmake option BUILD_TESTING to OFF
We have our own WITH_UNIT_TESTS option, and we call ENABLE_TESTING()
accordingly.  The option BUILD_TESTING comes from the CTest cmake
module, which does

option(BUILD_TESTING "..." ON)
if (BUILD_TESTING)
   # ...
   enable_testing()
   # ...
endif()

Several of our 3rd party libraries have cmake code which includes the
CTest module, which means that the value of BUILD_TESTING is
non-deterministic: it depends on what libraries are used as "bundled"
or "system".

To ensure consistency, disable BUILD_TESTING in the top level
CMakeLists.txt. We may want to build our own unit tests, but we do not
want to build whatever testing libraries/executables that come with
3rd party code.

Change-Id: If71ddbd0c1a291d6b4a68d9cd34c179cd28df82c