Bug#37324137: MySQL error when optimizing and updating table simultaneously.
Description:
------------
An error happens when concurrent optimize and update is done
on a table with virtual columns that are not part of any index
in that table. The error is caused by inconsistiency during
serialization and deserialization of values in undo log during
table rebuild. Values of virtual columns are not serialized to
undo log except ones that are part of some index. But when reading
from undo log during table rebuild it incorrectly tries to read
virtual column values from it that it shouldn't.

Fix:
----
There is no attempt to read values of virtual columns that
are not part of any index from undo log during full table rebuild.

Change-Id: I964f1f0bd482990d2af6a3529eb35f8228342e78