Bug#36716069 Compile MySQL with GCC 15 [5.7 noclose]
This patch is for 5.7, which is EOL, but we would still like to be
able to build it, to compare features/results in 5.7 vs newer
releases.

Disable -Wold-style-definition warnings a few places

Change sig_return typedef in my_global.h to fix:
my_lock.c:200:19: error: assignment to 'sig_return' ...
-Wincompatible-pointer-types

One more -Wincompatible-pointer-types fix in a unit test.

Change-Id: Ia59e1c4b8d24ba8e678fa92ff2a1c67dae115866