Bug#38105159: Glitch in handling of TIME values in NOT IN and range optimizer
Part 2: Implement class Item_cache_time for IN queries that involve
        TIME values.

The evaluation item used with the range optimizer used to be Item_int.
It is now replaced with Item_cache_time. This class is also extended
with a new function store_value() that makes it possible to store an
explicit TIME value in the cache object.

IN queries on the form <time_col> IN (<char_value>, <char_value> ...)
are now evaluated using In_vector_time instead of in_time_as_longlong.

IN queries on the form <time_col> IN (<time_value>, <time_value> ...)
are still evaluated using In_vector_string.

Change-Id: I9f08d830095b3bad4ba0dba5d093a8fb69015892