Bug#38099912 While restarting router, json file empty error.
When the Router discovers any changes that need to be reflected
in the dynamic state.json file it updates the file on the disk.
This is done directly on that file which can be risky if something
goes wrong while the file is been written to.

This patch changes the behaviour so that the Router first uses
state.json.tmp file for writing and only if that is successful,
renames it to the proper state.json.
This way if the writing fails the old file is preserved.

Change-Id: I67358a6ca1b6e9c8cad08d8384be6878ee4c087f